fix: melhora queries para supabase

+++ src/main/java/com/api/main/dto/UsuarioDTO.java
@@ -11,7 +11,7 @@ public class UsuarioDTO {
 
 	private String nome;
 
-	private Integer cpfCnpj;
+	private Long cpfCnpj;
 
 	private Set<DocumentoModel> documentos = new HashSet<>();
 	
@@ -54,11 +54,11 @@ public class UsuarioDTO {
 		this.nome = nome;
 	}
 
-	public Integer getCpfCnpj() {
+	public Long getCpfCnpj() {
 		return cpfCnpj;
 	}
 
-	public void setCpfCnpj(Integer cpfCnpj) {
+	public void setCpfCnpj(Long cpfCnpj) {
 		this.cpfCnpj = cpfCnpj;
 	}
